refactor: extract error alert into separate component
- Extract error display logic from AlertComp to AlertError component
- Create reusable AlertModal wrapper component using Svelte 5 Snippets
- Add section comments to AlertComp for better code organization
- Move stack trace translation logic and styles to AlertError
