+++ src/lib/Others/Alerts/AlertModal.svelte
@@ -1,0 +1,23 @@
+<script lang="ts">
+    import type { Snippet } from "svelte"
+
+    type Props = {
+        children: Snippet
+        title?: string
+        titleColor?: "red" | "green"
+    }
+
+    let { children, title, titleColor = "red" }: Props = $props()
+</script>
+
+<div class="absolute w-full h-full z-50 bg-black/50 flex justify-center items-center">
+    <div class="bg-darkbg p-4 break-any rounded-md flex flex-col max-w-3xl max-h-full overflow-y-auto">
+        {#if title}
+            <h2 class="mt-0 mb-2 w-40 max-w-full" class:text-red-700={titleColor === "red"} class:text-green-700={titleColor === "green"}>
+                {title}
+            </h2>
+        {/if}
+
+        {@render children()}
+    </div>
+</div>
